restructure pyproject.toml for plugins
